# doc/swagger.yaml
openapi: '3.0.0'
info:
  version: 1.0.0
  title: Vessels
  description: |
    The GFW Vessel API allows third parties to request aggregated information about vessels, such as identifiers, characteristics and tracks.
    This API requires an API key to use. The key should be included in the `Authorization` header of all requests to the API. Example: `Authorization: Bearer <token>`
servers:
  - url: https://gateway.api.dev.globalfishingwatch.org
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
  schemas:
    ValidationError:
      type: object
      required:
        - fields
        - general
      properties:
        fields:
          type: array
          items:
            type: object
            required:
              - field
              - errors
            properties:
              field:
                type: string
              errors:
                type: array
                items:
                  type: object
                  required:
                    - code
                    - description
                  properties:
                    code:
                      type: string
                    description:
                      type: string
        general:
          type: array
          items:
            type: object
            required:
              - code
              - description
            properties:
              code:
                type: string
              description:
                type: string
...
paths:
  '/tilesets/{tileset}/vessels':
    get:
      summary: |
        Searches for vessels
      tags:
        - VesselsAPI
      description: |
        Searches for a vessel given a free form query.The query will be matched against any identifiers for the vessel, and a paginated, ranked results list will be returned. This endpoint only returns vessels associated with a given tileset.
      parameters:
        - name: tileset
          in: path
          description: |
            Tileset to look the vessel from
          required: true
          schema:
            type: string
        - name: query
          in: query
          description: |
            Search query.
          required: true
          schema:
            type: string
        - name: queryFields
          in: query
          description: |
            List of fields to match with the search query instead of the entire object property
          schema:
            type: array
            items:
              type: string
        - name: querySuggestions
          in: query
          description: |
            Includes suggested results for the given query
          schema:
            type: boolean
            default: false
        - name: limit
          in: query
          description: |
            Amount of search results to return. Maximum 25.
          required: false
          schema:
            type: integer
            maximum: 25
            default: 10
        - name: offset
          in: query
          description: |
            Offset into the search results, used for pagination.
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: |
            Expected response to a successful request.
        '401':
          description: Unauthorized, either the api key is invalid or expired.
        '400':
          description: Bad request. Invalid or wrong parameters were provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  '/tilesets/{tileset}/vessels/{vesselId}':
    get:
      summary: |
        Obtains the characteristics of a single vessel
      description: |
        Obtains all the characteristics that describe a single vessel, such as its name and identifiers.
      tags:
        - VesselsAPI
      parameters:
        - name: tileset
          in: path
          description: |
            Tileset to query the vessel from
          required: true
          schema:
            type: string
        - name: vesselId
          in: path
          description: |
            Id of the vessel to get the characteristics from
          required: true
          schema:
            type: string
      responses:
        '200':
          description: |
            Expected response to a successful request.
        '401':
          description: Unauthorized, either the api key is invalid or expired.
        '400':
          description: Bad request. Invalid or wrong parameters were provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  '/datasets/{dataset}/vessels':
    get:
      summary: |
        Searches for vessels
      description: |
        Searches for a vessel given a free form query. The query will be matched against any identifiers for the vessel, and a paginated, ranked results list will be returned.
      tags:
        - VesselsAPI
      parameters:
        - name: dataset
          in: path
          description: |
            Dataset to query vessels from
          required: true
          schema:
            type: string
        - name: query
          in: query
          description: |
            Search query.
          required: true
          schema:
            type: string
        - name: queryFields
          in: query
          description: |
            List of fields to match with the search query instead of the entire object property
          schema:
            type: array
            items:
              type: string
        - name: querySuggestions
          in: query
          description: |
            Includes suggested results for the given query
          schema:
            type: boolean
            default: false
        - name: limit
          in: query
          description: |
            Amount of search results to return. Maximum 25.
          required: false
          schema:
            type: integer
            maximum: 25
            default: 10
        - name: offset
          in: query
          description: |
            Offset into the search results, used for pagination.
          required: false
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: |
            Expected response to a successful request.
        '401':
          description: Unauthorized, either the api key is invalid or expired.
        '400':
          description: Bad request. Invalid or wrong parameters were provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  '/datasets/{dataset}/vessels/{vesselId}':
    get:
      summary: |
        Obtains the characteristics of a single vessel
      description: |
        Obtains all the characteristics that describe a single vessel, such as its name and identifiers.
      tags:
        - VesselsAPI
      parameters:
        - name: dataset
          in: path
          description: |
            Dataset to query the vessel from
          required: true
          schema:
            type: string
        - name: vesselId
          in: path
          description: |
            Id of the vessel to get the characteristics from
          required: true
          schema:
            type: string
      responses:
        '200':
          description: |
            Expected response to a successful request.
        '401':
          description: Unauthorized, either the api key is invalid or expired.
        '400':
          description: Bad request. Invalid or wrong parameters were provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
...
  '/v1/vessels':
    get:
      summary: |
        List vessels
      description: |
        Lists vessels given a list of vessels id.
      tags:
        - VesselsAPI
      parameters:
        - name: datasets
          in: query
          description: |
            Dataset to query vessels from
          required: true
          schema:
            type: string
        - name: ids
          in: query
          description: |
            List of vessel id to get these vessels. The ids must be separated by comma.
          required: true
          schema:
            type: string
        - name: limit
          in: query
          description: |
            Amount of search results to return. Maximum 25.
          required: false
          schema:
            type: integer
            maximum: 25
            default: 10
        - name: offset
          in: query
          description: |
            Offset into the search results, used for pagination.
          required: false
          schema:
            type: integer
            default: 0
        - name: binary
          in: query
          description: |
            Return data in binary format.
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: |
            Expected response to a successful request.
        '401':
          description: Unauthorized, either the api key is invalid or expired.
        '400':
          description: Bad request. Invalid or wrong parameters were provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  '/v1/vessels/schema':
    get:
      summary: |
        Obtains the schema of the vessel
      description: |
        Obtains the schema of the vessel for each dataset
      tags:
        - VesselsAPI
      parameters:
        - name: datasets
          in: query
          description: |
            Dataset to query vessels from
          required: true
          schema:
            type: string
      responses:
        '200':
          description: |
            Expected response to a successful request.
        '401':
          description: Unauthorized, either the api key is invalid or expired.
        '400':
          description: Bad request. Invalid or wrong parameters were provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  '/v1/vessels/search':
    get:
      summary: |
        Search vessels
      description: |
        Searches for a vessel given a free form query. The query will be matched against any identifiers for the vessel, and a paginated, ranked results list will be returned.
      tags:
        - VesselsAPI
      parameters:
        - name: datasets
          in: query
          description: |
            Dataset to query vessels from
          required: true
          schema:
            type: string
        - name: query
          in: query
          description: |
            Search query.
          required: true
          schema:
            type: string
        - name: queryFields
          in: query
          description: |
            List of fields to match with the search query instead of the entire object property
          schema:
            type: array
            items:
              type: string
        - name: suggestField
          in: query
          description: |
            Includes suggested results for the given query for this property
          schema:
            type: string
            enum: [ shipname, imo, mmsi, flag ]
            default: shipname
        - name: limit
          in: query
          description: |
            Amount of search results to return. Maximum 25.
          required: false
          schema:
            type: integer
            maximum: 25
            default: 10
        - name: offset
          in: query
          description: |
            Offset into the search results, used for pagination.
          required: false
          schema:
            type: integer
            default: 0
        - name: binary
          in: query
          description: |
            Return data in binary format.
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: |
            Expected response to a successful request.
        '401':
          description: Unauthorized, either the api key is invalid or expired.
        '400':
          description: Bad request. Invalid or wrong parameters were provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  '/v1/vessels/advanced-search':
    get:
      summary: |
        Search vessels
      description: |
        Searches for a vessel given a query (SQL syntax).
      tags:
        - VesselsAPI
      parameters:
        - name: datasets
          in: query
          description: |
            Dataset to query vessels from
          required: true
          schema:
            type: string
        - name: query
          in: query
          description: |
            Search query.
          required: false
          schema:
            type: string
          example: shipname LIKE 'GUAP%' and mmsi = '265555280'
        - name: limit
          in: query
          description: |
            Amount of search results to return. Maximum 25.
          required: false
          schema:
            type: integer
            maximum: 25
            default: 10
        - name: offset
          in: query
          description: |
            Offset into the search results, used for pagination.
          required: false
          schema:
            type: integer
            default: 0
        - name: binary
          in: query
          description: |
            Return data in binary format.
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: |
            Expected response to a successful request.
        '401':
          description: Unauthorized, either the api key is invalid or expired.
        '400':
          description: Bad request. Invalid or wrong parameters were provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  '/v1/vessels/{vesselId}':
    get:
      summary: |
        Obtains the characteristics of a single vessel
      description: |
        Obtains all the characteristics that describe a single vessel, such as its name and identifiers.
      tags:
        - VesselsAPI
      parameters:
        - name: datasets
          in: query
          description: |
            Dataset to query vessels from
          required: true
          schema:
            type: string
        - name: vesselId
          in: path
          description: |
            Id of the vessel to get the characteristics from
          required: true
          schema:
            type: string
        - name: binary
          in: query
          description: |
            Return data in binary format.
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: |
            Expected response to a successful request.
        '401':
          description: Unauthorized, either the api key is invalid or expired.
        '400':
          description: Bad request. Invalid or wrong parameters were provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
...